chore: rename alerts folder
